reorganize layout components file structure
