--- src/components/layout/mobile/navbar/index.tsx
@@ -1,3 +1,0 @@
-import React from "react";
-import { AppBar, Toolbar } from "@material-ui/core";
-import {Logo} from "@content/as"
