+++ src/components/layout/navbar/expanded-mobile-navbar.tsx
