+++ src/pages/index.tsx
@@ -1,48 +1,11 @@
 import React from "react";
-import Button from "@components/form/Button";
-import Layout from "@components/layout";
-import SEO from "@components/layout/seo";
+import Layout, { SEO } from "@components/layout";
 
 const LandingPage = () => {
   return (
     <Layout>
       <SEO />
       <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3> <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      <h3>Reliable Manufacturer of JCB Parts</h3>
-      {/*  <Button rounded onClick={() => console.log("asdf")}>
-        Try it for free
-      </Button>
-      <h2>Landing Page</h2> */}
     </Layout>
   );
 };
